+++ tests/Unit/src/Tags/AudioTest.php
@@ -4,6 +4,7 @@ namespace umanskyi31\opengraph\test\Unit\src\Tags;
 use PHPUnit\Framework\TestCase;
 use umanskyi31\opengraph\Exceptions\OpenGraphException;
 use umanskyi31\opengraph\OpenGraph;
+use umanskyi31\opengraph\OpenGraphConfiguration;
 use umanskyi31\opengraph\Tags\Audio;
 
 class AudioTest extends TestCase
@@ -17,7 +18,7 @@ class AudioTest extends TestCase
     {
         parent::setUp();
 
-        $this->opengraph = new OpenGraph();
+        $this->opengraph = new OpenGraphConfiguration();
     }
 
     /**