+++ tests/Unit/src/Tags/BasicTest.php
@@ -3,6 +3,7 @@ namespace umanskyi31\opengraph\test\Unit\src\Tags;
 
 use PHPUnit\Framework\TestCase;
 use umanskyi31\opengraph\OpenGraph;
+use umanskyi31\opengraph\OpenGraphConfiguration;
 use umanskyi31\opengraph\Tags\Basic;
 
 class BasicTest extends TestCase
@@ -16,7 +17,7 @@ class BasicTest extends TestCase
     {
         parent::setUp();
 
-        $this->opengraph = new OpenGraph();
+        $this->opengraph = new OpenGraphConfiguration();
     }
 
     /**